# openapi.yaml
openapi: 3.0.3
info:
  title: API Pedidos e Gestão
  version: 1.0.0
  description: >
    API para gerenciamento de usuários, produtos, unidades e pedidos.
    Documentação gerada a partir dos arquivos de rota fornecidos (usuarios.js, produtos.js, unidades.js, pedidos.js).
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProdutoCreate:
      type: object
      required: [nome, preco, estoque]
      properties:
        nome:
          type: string
          example: "Laranja"
        preco:
          type: number
          format: float
          example: 5.0
        categoria:
          type: string
          example: "frutas"
        estoque:
          type: integer
          example: 70
    Produto:
      allOf:
        - $ref: '#/components/schemas/ProdutoCreate'
        - type: object
          properties:
            id:
              type: integer
              example: 2
    UnidadeCreate:
      type: object
      required: [nome, endereco, cep]
      properties:
        nome:
          type: string
          example: "Valadares"
        endereco:
          type: string
          example: "Av. Minas Gerais, 987"
        cep:
          type: string
          example: "11111-000"
    Unidade:
      allOf:
        - $ref: '#/components/schemas/UnidadeCreate'
        - type: object
          properties:
            id:
              type: integer
              example: 1
...
  responses:
...
    Forbidden:
      description: Acesso negado (role insuficiente)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Sem permissão para acessar este recurso"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Recurso não encontrado"
...
paths:
...
  /produtos:
    get:
      tags: [Produtos]
      summary: Listar produtos (público)
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
    post:
      tags: [Produtos]
      summary: Criar produto (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoCreate'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /produtos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Produtos]
      summary: Atualizar produto (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: { type: string }
                preco: { type: number }
                categoria: { type: string }
                estoque: { type: integer }
      responses:
        '200':
          description: Produto atualizado
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Produtos]
      summary: Remover produto (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Produto removido
        '404':
          $ref: '#/components/responses/NotFound'
  /unidades:
    get:
      tags: [Unidades]
      summary: Listar unidades
      responses:
        '200':
          description: Lista de unidades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Unidade'
    post:
      tags: [Unidades]
      summary: Criar unidade (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnidadeCreate'
      responses:
        '201':
          description: Unidade criada
        '403':
          $ref: '#/components/responses/Forbidden'
  /unidades/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Unidades]
      summary: Atualizar unidade (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: { type: string }
                endereco: { type: string }
                cep: { type: string }
      responses:
        '200':
          description: Unidade atualizada
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Unidades]
      summary: Remover unidade (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unidade removida
        '404':
          $ref: '#/components/responses/NotFound'
...
  /pedidos/relatorios/produtos:
    get:
      tags: [Relatorios]
      summary: Relatório de produtos mais consumidos (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ranking de produtos por receita
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalProdutos: { type: integer }
                  receitaGlobal: { type: number }
                  ranking:
                    type: array
                    items:
                      type: object
                      properties:
                        produtoId: { type: integer }
                        nome: { type: string }
                        quantidadeTotal: { type: integer }
                        receitaTotal: { type: number }
                        percentualReceita: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
  /pedidos/relatorios/unidades:
    get:
      tags: [Relatorios]
      summary: Relatório de vendas por unidade (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ranking por unidade
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUnidades: { type: integer }
                  receitaGlobal: { type: string }
                  ranking:
                    type: array
                    items:
                      type: object
                      properties:
                        unidadeId: { type: integer }
                        nome: { type: string }
                        totalPedidos: { type: integer }
                        receitaTotal: { type: string }
                        percentualReceita: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
...